fix(): fix CR request changes
