+++ src/app/app/ansyn/ansyn/ansyn.component.less
@@ -1,4 +1,4 @@
-@import "../../../../../node_modules/@ansyn/core/less/common";
+@import '~@ansyn/core/less/common';
 
 @keyframes pinned {
 	from {
